backend/balance_service/requirements.txt
asgiref==3.11.1
confluent-kafka==2.4.0
Django==5.0.14
djangorestframework==3.15.2
factory_boy==3.3.3
Faker==40.13.0
gunicorn==22.0.0
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
psycopg[binary]==3.1.20
Pygments==2.20.0
pytest==9.0.3
pytest-django==4.8.0
pytest-mock==3.14.1
python-decouple==3.8
sqlparse==0.5.5
typing_extensions==4.15.0
boto3==1.34.162
django-storages==1.14.4
setuptools==81.0.0
# Observability
django-prometheus==2.3.1
python-json-logger==2.0.7
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-django==0.45b0
opentelemetry-exporter-otlp-proto-grpc==1.24.0

